# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Store with promo codes for various stores",
        "title": "Freebie-shop",
        "contact": {
            "name": "Sergey Nekrasov",
            "url": "https://vk.com/serega_nekrasov",
            "email": "79508031750@yandex.ru"
        },
        "version": "1.0"
    },
    "host": "127.0.0.1:8080",
    "basePath": "/",
    "paths": {
        "/promos": {
            "get": {
                "description": "Get a list of all promos",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Info"
                ],
                "summary": "Get all records",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.PromosDocs"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/promos.PromoError"
                        }
                    }
                }
            }
        },
        "/promos/": {
            "post": {
                "description": "Adding a new promo to database",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Add"
                ],
                "summary": "Add a new promo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Магазин",
                        "name": "Store",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Скидка",
                        "name": "Discount",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Цена",
                        "name": "Price",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Количество",
                        "name": "Quantity",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "description": "Промокоды(запись в виде массива)",
                        "name": "Promo",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/promos.PromoCreated"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/promos.PromoError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/promos.PromoError"
                        }
                    }
                }
            }
        },
        "/promos/:uuid": {
            "get": {
                "description": "Get the price using the promo uuid",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Info"
                ],
                "summary": "Get price for a promo",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "UUID промо",
                        "name": "UUID",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/promos.PromoPrice"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/promos.PromoError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/promos.PromoError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/promos.PromoError"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a promo using its uuid",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Delete"
                ],
                "summary": "Delete a promo",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "UUID промо",
                        "name": "UUID",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/promos.PromoDeleted"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/promos.PromoError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/promos.PromoError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/promos.PromoError"
                        }
                    }
                }
            }
        },
        "/promos/:uuid/": {
            "put": {
                "description": "Change the promo price using its uuid",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Change"
                ],
                "summary": "Change promo price",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "UUID промо",
                        "name": "UUID",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Новая цена",
                        "name": "Price",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/promos.PromoChanged"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/promos.PromoError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/promos.PromoError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/promos.PromoError"
                        }
                    }
                }
            }
        },
        "/promos/random": {
            "post": {
                "description": "Adding a new random promo to database",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Add"
                ],
                "summary": "Add a new random promo",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Количество",
                        "name": "Quantity",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/promos.PromoCreated"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/promos.PromoError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/promos.PromoError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.PromosDocs": {
            "type": "object",
            "properties": {
                "discount": {
                    "type": "integer"
                },
                "price": {
                    "type": "integer"
                },
                "promo": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "quantity": {
                    "type": "integer"
                },
                "store": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
        "promos.PromoChanged": {
            "type": "object",
            "properties": {
                "changed": {
                    "type": "boolean"
                }
            }
        },
        "promos.PromoCreated": {
            "type": "object",
            "properties": {
                "created": {
                    "description": "success",
                    "type": "boolean"
                }
            }
        },
        "promos.PromoDeleted": {
            "type": "object",
            "properties": {
                "deleted": {
                    "type": "boolean"
                }
            }
        },
        "promos.PromoError": {
            "type": "object",
            "properties": {
                "description": {
                    "description": "description",
                    "type": "string"
                },
                "error": {
                    "description": "error",
                    "type": "string"
                },
                "type": {
                    "description": "type",
                    "type": "string"
                }
            }
        },
        "promos.PromoPrice": {
            "type": "object",
            "properties": {
                "price": {
                    "type": "integer"
                }
            }
        }
    }
}